+++ sms/tests/test_sms.py
@@ -171,3 +171,8 @@ def test_submit_sms_conversion_402():
         sms.submit_sms_conversion('3295d748-4e14-4681-af78-166dca3c5aab')
     except HttpRequestError as err:
         assert err.message == '402 response from https://api.nexmo.com/conversions/sms.'
+
+
+def test_http_client_property():
+    sms = Sms(HttpClient(Auth(api_key='qwerasdf', api_secret='1234qwerasdfzxcv')))
+    assert isinstance(sms.http_client, HttpClient)