+++ voice/src/vonage_voice/models/__init__.py
@@ -1,4 +1,4 @@
-from .common import AdvancedMachineDetection
+from .common import AdvancedMachineDetection, Phone, Sip, Vbc, Websocket
 from .connect_endpoints import (
     AppEndpoint,
     OnAnswer,
@@ -7,39 +7,67 @@ from .connect_endpoints import (
     VbcEndpoint,
     WebsocketEndpoint,
 )
-from .enums import Channel, ConnectEndpointType, NccoActionType
+from .enums import (
+    CallState,
+    Channel,
+    ConnectEndpointType,
+    NccoActionType,
+    TtsLanguageCode,
+)
 from .input_types import Dtmf, Speech
 from .ncco import Connect, Conversation, Input, NccoAction, Notify, Record, Stream, Talk
-from .requests import CreateCallRequest, Phone, Sip, ToPhone, Vbc, Websocket
-from .responses import CallStatus, CreateCallResponse
+from .requests import (
+    AudioStreamOptions,
+    CreateCallRequest,
+    ListCallsFilter,
+    ToPhone,
+    TtsStreamOptions,
+)
+from .responses import (
+    CallInfo,
+    CallList,
+    CallMessage,
+    CreateCallResponse,
+    Embedded,
+    Links,
+)
 
 __all__ = [
     'AdvancedMachineDetection',
-    'CreateCallRequest',
-    'ToPhone',
-    'Sip',
-    'Websocket',
-    'Vbc',
-    'Phone',
-    'NccoAction',
-    'Channel',
-    'NccoActionType',
-    'ConnectEndpointType',
-    'OnAnswer',
-    'PhoneEndpoint',
     'AppEndpoint',
-    'WebsocketEndpoint',
-    'SipEndpoint',
-    'VbcEndpoint',
-    'Dtmf',
-    'Speech',
-    'CreateCallResponse',
-    'CallStatus',
-    'Record',
-    'Conversation',
+    'AudioStreamOptions',
+    'CallInfo',
+    'CallList',
+    'CallMessage',
+    'CallState',
+    'Channel',
     'Connect',
-    'Talk',
-    'Stream',
+    'ConnectEndpointType',
+    'Conversation',
+    'CreateCallRequest',
+    'CreateCallResponse',
+    'Dtmf',
+    'Embedded',
     'Input',
+    'ListCallsFilter',
+    'Links',
+    'NccoAction',
+    'NccoActionType',
     'Notify',
+    'OnAnswer',
+    'Phone',
+    'PhoneEndpoint',
+    'Record',
+    'Sip',
+    'SipEndpoint',
+    'Speech',
+    'Stream',
+    'Talk',
+    'ToPhone',
+    'TtsLanguageCode',
+    'TtsStreamOptions',
+    'Vbc',
+    'VbcEndpoint',
+    'Websocket',
+    'WebsocketEndpoint',
 ]