+++ vonage_utils/src/vonage_utils/types.py
@@ -1,6 +1,5 @@
-from typing import Annotated
-
 from pydantic import Field
+from typing_extensions import Annotated
 
 PhoneNumber = Annotated[str, Field(pattern=r'^[1-9]\d{6,14}$')]
 Dtmf = Annotated[str, Field(pattern=r'^[0-9#*p]+$')]