+++ sms/CHANGES.md
@@ -1,3 +1,6 @@
+# 1.1.0
+- Add `http_client` property
+
 # 1.0.2
 - Internal refactoring