+++ verify_v2/CHANGES.md
@@ -1,2 +1,5 @@
+# 1.1.0
+- Add `http_client` property
+
 # 1.0.0
 - Initial upload