+++ vonage_utils/CHANGES.md
@@ -1,3 +1,8 @@
+# 1.1.0
+- Add `Dtmf` and `SipUri` types
+- Add `Link` model
+- Internal refactoring
+
 # 1.0.1
 - Add `PhoneNumber` type